+++ packages/mantine/src/components/prompt/Prompt.tsx
@@ -47,6 +47,7 @@ interface PromptType {
 }
 
 export const Prompt: PromptType = ({children, variant = 'info', ...otherProps}) => {
+    const {classNames: classesProps, ...otherPropsWithoutClasses} = otherProps;
     const {classes, cx} = useStyles();
     const convertedChildren = Children.toArray(children) as ReactElement[];
 
@@ -62,7 +63,13 @@ export const Prompt: PromptType = ({children, variant = 'info', ...otherProps}) 
     };
 
     return (
-        <Modal variant="prompt" padding={0} classNames={classNames} size={'sm'} {...otherProps}>
+        <Modal
+            variant="prompt"
+            padding={0}
+            classNames={{...classNames, ...classesProps}}
+            size={'sm'}
+            {...otherPropsWithoutClasses}
+        >
             <div className={classes.innerBody}>{otherChildren}</div>
             {footer}
         </Modal>